Fix submarine -> kogito rename leftovers
