+++ openshift-example-tests/src/test/java/org/kogito/examples/openshift/JbpmQuarkusExampleIntegrationTest.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.submarine.examples.openshift;
+package org.kogito.examples.openshift;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -23,24 +23,24 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.submarine.examples.openshift.Project;
-import org.submarine.examples.openshift.TestConfig;
-import org.submarine.examples.openshift.deployment.Deployer;
-import org.submarine.examples.openshift.deployment.HttpDeployment;
+import org.kogito.examples.openshift.Project;
+import org.kogito.examples.openshift.TestConfig;
+import org.kogito.examples.openshift.deployment.Deployer;
+import org.kogito.examples.openshift.deployment.HttpDeployment;
 
-public class JbpmSpringBootExampleIntegrationTest {
+public class JbpmQuarkusExampleIntegrationTest {
 
     private static Project project;
     private static HttpDeployment kaasDeloyment;
 
     @BeforeClass
     public static void setUp() throws MalformedURLException {
-        URL assetsUrl = new URL("https://github.com/kiegroup/submarine-examples");
-        String gitContextDir = "jbpm-springboot-example";
+        URL assetsUrl = new URL("https://github.com/kiegroup/kogito-examples");
+        String gitContextDir = "jbpm-quarkus-example";
 
         String randomProjectName = RandomStringUtils.randomAlphanumeric(4).toLowerCase();
         project = Project.create("jbpm-example-" + randomProjectName);
-        kaasDeloyment = Deployer.deployKaasUsingS2iAndWait(project, assetsUrl, gitContextDir, TestConfig.getKaasS2iSpringBootBuilderImage(), TestConfig.getKaasSpringBootRuntimeImage());
+        kaasDeloyment = Deployer.deployKaasUsingS2iAndWait(project, assetsUrl, gitContextDir, TestConfig.getKaasS2iQuarkusBuilderImage(), TestConfig.getKaasQuarkusRuntimeImage());
     }
 
     @AfterClass
